UBUNTU: SAUCE: regulator: sun20i: Use device_node_to_regmap()
BugLink: https://bugs.launchpad.net/bugs/1975592

Using syscon_node_to_regmap() on the parent node will also grab
the clock and resets defined in the parent node. But there is already a
driver doing that, so this will result in an error.
Fix it by using device_node_to_regmap() which will not grab those
resources.

Fixes: 871ec997fa3e ("regulator: sun20i: Add support for Allwinner D1 LDOs")
Signed-off-by: Emil Renner Berthing <emil.renner.berthing@canonical.com>