FROMLIST: dt-bindings: riscv: Add B ISA extension description
Add description of the single-letter "B" extennsion for Bit Manipulation.
B is mandatory for RVA23U64.

The B extension is ratified in the 20240411 version of the unprivileged
ISA specification. According to the ratified spec, "the B standard
extension comprises instructions provided by the Zba, Zbb, and Zbs
extensions.

Hence add a schema check rule to enforce that B implies Zba, Zbb and Zbs.

Signed-off-by: Guodong Xu <guodong@riscstar.com>
Link: https://lore.kernel.org/r/20251222-k3-basic-dt-v2-7-3af3f3cd0f8a@riscstar.com
Signed-off-by: Han Gao <gaohan@iscas.ac.cn>